:recycle: to separate components of BestCombinationPickPage
