+++ src/components/BestCombinationPick/NavigateListPageButton.tsx
@@ -1,0 +1,32 @@
+import { useNavigate } from 'react-router-dom';
+import Button from '@components/UI/Button';
+import styled from '@emotion/styled';
+import { changeRem } from '@styles/mixin';
+import { 인터페이스_꿀조합선택페이지_필터 } from '../../types/ISandwich';
+
+function NavigateListPageButton({ filteredAttr }: { filteredAttr: 인터페이스_꿀조합선택페이지_필터 }) {
+  const navigate = useNavigate();
+
+  const 꿀조합_목록_페이지로_이동하기 = () => {
+    navigate('/best-combination', { state: filteredAttr });
+  };
+
+  return (
+    <ButtonWrap>
+      <Button
+        onClick={꿀조합_목록_페이지로_이동하기}
+        designType="primaryGreen"
+        width={changeRem(330)}
+        height={changeRem(50)}>
+        꿀 조합 보러가기
+      </Button>
+    </ButtonWrap>
+  );
+}
+
+export default NavigateListPageButton;
+
+const ButtonWrap = styled.div`
+  position: absolute;
+  bottom: 120px;
+`;
